+++ crates/feather/src/internals/mod.rs
@@ -2,3 +2,6 @@ mod app;
 mod context;
 pub use app::App;
 pub use context::AppContext;
+mod error_stack;
+
+pub use feather_runtime::{Method,Uri,HeaderMap,HeaderName,HeaderValue};