+++ crates/feather/src/lib.rs
@@ -3,12 +3,25 @@
 /// The [`Middleware`] trait and some common middleware primitives.
 pub mod middleware;
 
-/// Synchronous API for Feather.
-mod internals;
+pub mod internals;
 #[cfg(feature = "jwt")]
 pub mod jwt;
 
-pub use feather_runtime::http::Request;
-pub use feather_runtime::http::Response;
+use std::error::Error;
+
+pub use feather_runtime::http::{Request,Response};
 pub use internals::{App, AppContext};
-pub use middleware::MiddlewareResult;
+pub use crate::middleware::MiddlewareResult;
+
+/// This is just a type alias for `Result<MiddlewareResult, Box<dyn Error>>;`  
+/// Outcome is used in All middlewares as a return type.
+pub type Outcome = Result<MiddlewareResult, Box<dyn Error>>;
+
+/// This macro is just a syntactic sugar over the `Ok(MiddlewareResult::Next)`
+/// syntax just to clear some Boilerplate
+#[macro_export]
+macro_rules! next {
+    () => {
+        Ok($crate::middleware::MiddlewareResult::Next)
+    };
+}