+++ crates/feather/src/middleware/common.rs
@@ -1,4 +1,4 @@
-use crate::internals::AppContext;
+use crate::{internals::AppContext, Outcome};
 use feather_runtime::http::{Request, Response};
 
 pub trait Middleware {
@@ -8,9 +8,10 @@ pub trait Middleware {
         request: &mut Request,
         response: &mut Response,
         ctx: &mut AppContext,
-    ) -> MiddlewareResult;
+    ) -> Outcome;
 }
 
+#[derive(Debug)]
 pub enum MiddlewareResult {
     /// Continue to the next middleware.
     Next,
@@ -25,27 +26,27 @@ impl Middleware for [&Box<dyn Middleware>] {
         request: &mut Request,
         response: &mut Response,
         ctx: &mut AppContext,
-    ) -> MiddlewareResult {
+    ) -> Outcome {
         for middleware in self {
             if matches!(
                 middleware.handle(request, response, ctx),
-                MiddlewareResult::NextRoute
+                Ok(MiddlewareResult::NextRoute)
             ) {
-                return MiddlewareResult::NextRoute;
+                return Ok(MiddlewareResult::NextRoute);
             }
         }
-        MiddlewareResult::Next
+        Ok(MiddlewareResult::Next)
     }
 }
 
 ///Implement the `Middleware` trait for Closures with Request and Response Parameters.
-impl<F: Fn(&mut Request, &mut Response, &mut AppContext) -> MiddlewareResult> Middleware for F {
+impl<F: Fn(&mut Request, &mut Response, &mut AppContext) -> Outcome> Middleware for F {
     fn handle(
         &self,
         request: &mut Request,
         response: &mut Response,
         ctx: &mut AppContext,
-    ) -> MiddlewareResult {
+    ) -> Outcome {
         self(request, response, ctx)
     }
 }
@@ -53,8 +54,7 @@ impl<F: Fn(&mut Request, &mut Response, &mut AppContext) -> MiddlewareResult> Mi
 /// Can be used to chain two middlewares together.
 /// The first middleware will be executed first.
 /// If it returns `MiddlewareResult::Next`, the second middleware will be executed.
-fn _chainer<A, B>(a: A, b: B) -> impl Middleware
-// Nvm the warning this is used in the macro
+pub fn _chainer<A, B>(a: A, b: B) -> impl Middleware
 where
     A: Middleware,
     B: Middleware,
@@ -62,10 +62,11 @@ where
     move |request: &mut Request,
           response: &mut Response,
           ctx: &mut AppContext|
-          -> MiddlewareResult {
+          -> Outcome {
         match a.handle(request, response, ctx) {
-            MiddlewareResult::Next => b.handle(request, response, ctx),
-            MiddlewareResult::NextRoute => MiddlewareResult::NextRoute,
+            Ok(MiddlewareResult::Next) => b.handle(request, response, ctx),
+            Ok(MiddlewareResult::NextRoute) => Ok(MiddlewareResult::NextRoute),
+            Err(e) => Err(e),
         }
     }
 }
@@ -76,7 +77,7 @@ where
 macro_rules! chain {
     ($first:expr, $($rest:expr),+ $(,)?) => {{
         let chained = $first;
-        $(let chained = $crate::chainer(chained, $rest);)+
+        $(let chained = $crate::middleware::common::_chainer(chained, $rest);)+
         chained
     }};
 }