+++ examples/basic-app/src/main.rs
@@ -1,6 +1,6 @@
 // Import dependencies from Feather
 use feather::middleware::builtins;
-use feather::{App, AppContext, MiddlewareResult};
+use feather::{App, AppContext, next};
 use feather::{Request, Response};
 
 // Main function - no async here!
@@ -13,11 +13,11 @@ fn main() {
         "/",
         |_request: &mut Request, response: &mut Response, _ctx: &mut AppContext| {
             response.send_text("Hello, world!");
-            MiddlewareResult::Next
+            next!()
         },
     );
     // Use the Logger middleware for all routes
     app.use_middleware(builtins::Logger);
-    // Listen on port 3000
-    app.listen("127.0.0.1:3000");
+    // Listen on port 5050
+    app.listen("127.0.0.1:5050");
 }