+++ examples/context/src/main.rs
@@ -1,6 +1,6 @@
 /// Use of the AppContext State Managment with Sqlite
 // Import Our Dependencies
-use feather::{App, AppContext, MiddlewareResult, Request, Response};
+use feather::{next, App, AppContext, Outcome, Request, Response};
 use rusqlite::{Connection, Result};
 use serde_json::json;
 
@@ -17,9 +17,8 @@ fn main() -> Result<()> {
         name  TEXT NOT NULL  
     )",
         [],
-    )
-    .unwrap(); // For the sake of this example we are going to unwrap where we can but on production code error should be handled
-    app.context().set_state(conn); // Store the conn inside of our context
+    )?;
+    app.context().set_state(conn); // Store the connection inside of our context
     // from now on context is only accesible inside the context
 
     app.post("/login", login);
@@ -30,21 +29,28 @@ fn main() -> Result<()> {
     Ok(())
 }
 // Post Route for loging in users
-fn login(req: &mut Request, res: &mut Response, ctx: &mut AppContext) -> MiddlewareResult {
+fn login(req: &mut Request, res: &mut Response, ctx: &mut AppContext) -> Outcome {
     let data = match req.json() {
         Ok(json) => json,
         Err(_) => {
-            res.status(400).send_json(json!({"error": "Invalid JSON"}));
-            return MiddlewareResult::Next;
+            res.set_status(400).send_json(json!({"error": "Invalid JSON"}));
+            return next!()
         }
     };
     println!("Received Json: {data}"); // Log it to see what we got
-    let db = ctx.get_state::<Connection>().unwrap(); // Get the connection from the context, remember we are not opening new connections here
-    let username = data.get("username").unwrap().to_string(); // Get the username
+    let db = ctx.get_state::<Connection>().unwrap(); // Get the Connection from the context. Unwrap is safe here because we know we set it before
+    let username = match data.get("username") {
+        Some(v) => v.to_string(),
+        None => {
+            res.set_status(400).send_text("No username found in the data!");
+            return next!()
+        }
+    };
+
     // Now Lets put it inside of our DB
     match db.execute("INSERT INTO person (name) VALUES (?1)", [username]) {
         //If it succeeds we send a successs message with 200 Code
-        Ok(rows_changed) => res.status(200).send_json(json!
+        Ok(rows_changed) => res.set_status(200).send_json(json!
         (
             {
                 "success":true,
@@ -53,7 +59,7 @@ fn login(req: &mut Request, res: &mut Response, ctx: &mut AppContext) -> Middlew
         )),
         //If it fails we send a successs message with 500 Code
         Err(e) => {
-            res.status(500).send_json(json!
+            res.set_status(500).send_json(json!
             (
                 {
                     "success":false,
@@ -62,18 +68,20 @@ fn login(req: &mut Request, res: &mut Response, ctx: &mut AppContext) -> Middlew
             println!("{e}")
         }
     };
-    feather::MiddlewareResult::Next
+    next!()
 }
 // Get Route for listing users
-fn get_user(_req: &mut Request, res: &mut Response, ctx: &mut AppContext) -> MiddlewareResult {
+fn get_user(_req: &mut Request, res: &mut Response, ctx: &mut AppContext) -> Outcome {
     let db = ctx.get_state::<Connection>().unwrap(); // Again Take our Connection from the context. that is still a single connection
-    let mut stmt = db.prepare("SELECT name FROM person").unwrap();
+    
+    // We can use the ? operator here because we are inside of a function that returns a Result
+    // We prepare a statement to select all names from the person table
+    let mut stmt = db.prepare("SELECT name FROM person")?;
     let users = stmt
-        .query_map([], |row| row.get::<_, String>(0))
-        .unwrap()
+        .query_map([], |row| row.get::<_, String>(0))?
         .filter_map(Result::ok)
         .collect::<Vec<_>>();
 
-    res.status(200).send_json(json!({ "users": users }));
-    MiddlewareResult::Next
+    res.set_status(200).send_json(json!({ "users": users }));
+    next!()
 }