+++ examples/middleware/src/main.rs
@@ -1,55 +1,37 @@
-// Import dependencies from Feather
-use feather::{App, AppContext, Request, Response};
-// Import the Middleware trait and some common middleware primitives
-use feather::middleware::builtins;
-use feather::middleware::{Middleware, MiddlewareResult};
-// Implementors of the Middleware trait are middleware that can be used in a Feather app.
-struct Custom;
-
-// The Middleware trait defines a single method `handle`,
-// which can mutate the request and response objects, then return a `MiddlewareResult`.
-impl Middleware for Custom {
-    fn handle(
-        &self,
-        request: &mut Request,
-        _response: &mut Response,
-        _ctx: &mut AppContext,
-    ) -> MiddlewareResult {
-        // Do stuff here
-        println!("Now running some custom middleware (struct Custom)!");
-        println!("And there's a request with path: {:?}", request.uri);
-        // and then continue to the next middleware in the chain
-        MiddlewareResult::Next
-    }
-}
-
-fn main() {
-    // Create a new instance of App
+use feather::{chain, middleware::builtins, next, App, AppContext, Outcome, Request, Response};
+mod middleware;
+use middleware::MyMiddleware;
+fn main(){
     let mut app = App::new();
 
-    // Use the builtin Logger middleware for all routes
-    app.use_middleware(builtins::Logger);
+    app.use_middleware(builtins::Logger);// We can easily use middlewares using this syntax
+    // We can also put Closures as a middleware parameter. that what makes Feather "Middleware-First"
+    app.use_middleware(|_req: &mut feather::Request,_res: &mut feather::Response,_ctx: &mut feather::AppContext|{
+        println!("Ow a Request: I am a Closure Middleware BTW");
+        next!()
+    }); 
+    app.use_middleware(MyMiddleware("Secret Codee".to_string()));
 
-    // Use the Custom middleware for all routes
-    app.use_middleware(Custom);
+    app.get("/", |_req: &mut Request,res: &mut Response,_ctx: &mut AppContext|{
+        res.send_text("Hellooo Feather Middleware Example");
+        res.set_status(200);
+        next!()
+    });
+    // You can also chain middlewares using the `chain!` macro 
+    // the first given middleware will always run first!
+    // You can also chain more than 2 middlewares
+    app.get("/chain", chain!(first,second));
+    
+    app.listen("127.0.0.1:5050");
+}
 
-    // Use another middleware defined by a function for all routes
-    app.use_middleware(
-        |_request: &mut Request, _response: &mut Response, _ctx: &mut AppContext| {
-            println!("Now running some custom middleware (closure)!");
-            MiddlewareResult::Next
-        },
-    );
-
-    // Define a route
-    app.get(
-        "/",
-        |_request: &mut Request, response: &mut Response, _ctx: &mut AppContext| {
-            response.send_text("Hello, world!");
-            MiddlewareResult::Next
-        },
-    );
-
-    // Listen on port 3000
-    app.listen("127.0.0.1:3000");
+fn first(_req:&mut Request,res:&mut Response, _ctx: &mut AppContext) -> Outcome{
+    println!("First Middleware Ran");
+    res.set_status(201);
+    next!()
+}
+fn second(_req:&mut Request,res:&mut Response, _ctx: &mut AppContext) -> Outcome{
+    println!("Second Ran");
+    res.send_text("Yep Chained middlewares");
+    next!()
 }